requirements.txt
python-gitlab==3.1.0
giturlparse==0.10.0